refactor: movie 관련 컴포넌트 분류
